+++ mq_radio/engine/__init__.py
@@ -1,4 +1,5 @@
 from mq_radio.engine.audio_devices import list_audio_devices
+from mq_radio.engine import au_insert
 from mq_radio.engine.audio_router import (
     AudioRouter,
     apply_audio_route_from_settings,
@@ -14,6 +15,7 @@ __all__ = [
     "MockEngine",
     "LiquidsoapEngine",
     "list_audio_devices",
+    "au_insert",
     "AudioRouter",
     "get_audio_router",
     "apply_audio_route_from_settings",